[Resource] [Rbac] Moved Rbac integration from ResourceBundle to RbacBundle
