+++ src/Sylius/Bundle/RbacBundle/DependencyInjection/SyliusRbacExtension.php
@@ -62,6 +62,15 @@ class SyliusRbacExtension extends AbstractResourceExtension implements PrependEx
      */
     public function prepend(ContainerBuilder $container)
     {
+        $this->prependDoctrineCache($container);
+        $this->prependSyliusResource($container);
+    }
+
+    /**
+     * @param ContainerBuilder $container
+     */
+    private function prependDoctrineCache(ContainerBuilder $container)
+    {
         if (!$container->hasExtension('doctrine_cache')) {
             throw new \RuntimeException('DoctrineCacheBundle must be registered!');
         }
@@ -71,5 +80,19 @@ class SyliusRbacExtension extends AbstractResourceExtension implements PrependEx
                 'sylius_rbac' => '%sylius.cache%',
             ],
         ]);
+    }
+    /**
+     * @param ContainerBuilder $container
+     */
+    private function prependSyliusResource(ContainerBuilder $container)
+    {
+        if (!$container->hasExtension('sylius_resource')) {
+            return;
+        }
+
+        $loader = new XmlFileLoader($container, new FileLocator(__DIR__.'/../Resources/config'));
+        $loader->load('resource_integration.xml');
+
+        $container->setAlias('sylius.resource_controller.authorization_checker', 'sylius.resource_controller.authorization_checker.rbac');
     }
 }
