+++ src/Sylius/Bundle/ResourceBundle/composer.json
@@ -51,7 +51,6 @@
         "phpspec/phpspec": "^2.4",
         "phpunit/phpunit": "^4.8",
         "sensio/generator-bundle": "^3.0",
-        "sylius/rbac-bundle": "^0.18",
         "twig/twig": "~1.11"
     },
     "suggest": {
